# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"InfraStemAPI","description":"","license":{"name":""},"version":"1.0.0"},"paths":{"/api/gamble":{"post":{"tags":[],"operationId":"gamble","requestBody":{"description":"Type of gamble","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Gamble"}}},"required":true},"responses":{"200":{"description":"Generates a list of numbers from 1 to 7","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GambleResults"}}}},"401":{"description":"Unauthorized - missing or invalid token"}},"security":[{"bearerAuth":[]}]}},"/api/post":{"get":{"tags":["post"],"description":"Retrieves 5 posts","operationId":"retrieve_posts","parameters":[{"name":"type","in":"path","required":true,"schema":{"$ref":"#/components/schemas/RetrieveBy"}},{"name":"page","in":"path","required":true,"schema":{"type":"integer","minimum":0}}],"responses":{"200":{"description":"Retrieves posts","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponse_KamerlinkPost"}}}},"401":{"description":"Unauthorized - missing or invalid token"}}},"post":{"tags":["post"],"operationId":"create_post","requestBody":{"description":"Creates a post","content":{"multipart/form-data":{"schema":{"type":"object","description":"\n\nPost request (creating a post)\n\n","required":["title","message","images"],"properties":{"images":{"type":"array","items":{"type":"array","items":{"type":"integer","format":"int32","minimum":0}},"contentMediaType":"application/octet-stream"},"message":{"type":"string"},"title":{"type":"string"}}}}},"required":true},"responses":{"200":{"description":"creates post","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostResponse"}}}},"401":{"description":"Unauthorized - missing or invalid token"}},"security":[{"bearerAuth":[]}]}},"/api/post/comment":{"get":{"tags":["post::comment"],"description":"Retrieves 5 posts","operationId":"retrieve_comments","parameters":[{"name":"type","in":"path","required":true,"schema":{"$ref":"#/components/schemas/RetrieveBy"}},{"name":"page","in":"path","required":true,"schema":{"type":"integer","minimum":0}},{"name":"post_id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Retrieves posts","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponse_Comment"}}}},"401":{"description":"Unauthorized - missing or invalid token"}}},"post":{"tags":["post::comment"],"operationId":"create_comment","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDraft"}}},"required":true},"responses":{"200":{"description":"Adds a comment to a post (returns a comment id)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Comment"}}}},"401":{"description":"Unauthorized - missing or invalid token"}},"security":[{"bearerAuth":[]}]}},"/api/post/comment/like":{"post":{"tags":["post::comment"],"operationId":"like_comment","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenericLike"}}},"required":true},"responses":{"200":{"description":"Toggles the like/unlike under a comment","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseGenericLike"}}}},"401":{"description":"Unauthorized - missing or invalid token"}},"security":[{"bearerAuth":[]}]}},"/api/post/comment/reply":{"post":{"tags":["post::comment"],"operationId":"add_reply_to_comment","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReplyDraft"}}},"required":true},"responses":{"200":{"description":"Adds a reply to a comment (returns the updated comment)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reply"}}}},"401":{"description":"Unauthorized - missing or invalid token"}},"security":[{"bearerAuth":[]}]}},"/api/post/like":{"post":{"tags":["post::like"],"operationId":"like_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LikePost"}}},"required":true},"responses":{"200":{"description":"Toggles the like/unlike under a post","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseLikePost"}}}},"401":{"description":"Unauthorized - missing or invalid token"}},"security":[{"bearerAuth":[]}]}},"/api/post/points":{"get":{"tags":["post::points"],"operationId":"check_points","parameters":[{"name":"post_id","in":"query","description":"Type of retrieval","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Returns how many points a specific user has spent on a post","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GivenPoints"}}}},"401":{"description":"Unauthorized - missing or invalid token"}},"security":[{"bearerAuth":[]}]},"post":{"tags":["post::points"],"operationId":"spend_points","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GivePoints"}}},"required":true},"responses":{"200":{"description":"Gives Points to a specific post","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseGivePoints"}}}},"401":{"description":"Unauthorized - missing or invalid token"}},"security":[{"bearerAuth":[]}]}},"/api/post/tags":{"post":{"tags":["post::tags"],"operationId":"retrieve_tags","parameters":[{"name":"type","in":"path","required":true,"schema":{"$ref":"#/components/schemas/RetrieveBy"}},{"name":"page","in":"path","required":true,"schema":{"type":"integer","minimum":0}}],"responses":{"200":{"description":"retrieves suggested","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponse_PostTag"}}}},"401":{"description":"Unauthorized - missing or invalid token"}},"security":[{"bearerAuth":[]}]}},"/api/user":{"get":{"tags":["user"],"description":"Retrieves 5 users","operationId":"retrieve_users","parameters":[{"name":"type","in":"path","required":true,"schema":{"$ref":"#/components/schemas/RetrieveBy"}},{"name":"page","in":"path","required":true,"schema":{"type":"integer","minimum":0}}],"responses":{"200":{"description":"Retrieves users","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponse_UserInfo"}}}},"401":{"description":"Unauthorized - missing or invalid token"}}}},"/api/user/{user_id}/posts":{"get":{"tags":["user"],"description":"Retrieves 5 users","operationId":"retrieve_user_posts","parameters":[{"name":"user_id","in":"path","description":"The id of the user to get the posts from","required":true,"schema":{"type":"string"}},{"name":"type","in":"path","required":true,"schema":{"$ref":"#/components/schemas/RetrieveBy"}},{"name":"page","in":"path","required":true,"schema":{"type":"integer","minimum":0}}],"responses":{"200":{"description":"Retrieves users","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginatedResponse_KamerlinkPost"}}}},"401":{"description":"Unauthorized - missing or invalid token"}}}}},"components":{"schemas":{"Comment":{"type":"object","required":["_id","user_id","post_id","message","created_at","likes","replies"],"properties":{"_id":{"$ref":"#/components/schemas/ObjectIdSchema"},"created_at":{"type":"string"},"likes":{"type":"integer","minimum":0},"message":{"type":"string"},"post_id":{"type":"string"},"replies":{"type":"array","items":{"$ref":"#/components/schemas/Reply"}},"user_id":{"type":"string"}}},"CommentDraft":{"type":"object","required":["post_id","message"],"properties":{"message":{"type":"string"},"post_id":{"type":"string"}}},"Gamble":{"type":"object","required":["gamble_type"],"properties":{"gamble_type":{"$ref":"#/components/schemas/GambleTypes"}}},"GambleResults":{"type":"object","properties":{"slots":{"type":["array","null"],"items":{"type":"integer","format":"int32","minimum":0}}}},"GambleTypes":{"type":"string","enum":["Slots"]},"GenericLike":{"type":"object","required":["_id"],"properties":{"_id":{"type":"string"}}},"GivePoints":{"type":"object","required":["points","post_id"],"properties":{"points":{"type":"integer","format":"int64"},"post_id":{"type":"string"}}},"GivenPoints":{"type":"object","required":["points_given","limit"],"properties":{"limit":{"type":"integer","format":"int64"},"points_given":{"type":"integer","format":"int64"}}},"KamerlinkError":{"type":"string","enum":[]},"KamerlinkPost":{"type":"object","required":["_id","user_id","created_at","title","message","img_urls","likes","points"],"properties":{"_id":{"$ref":"#/components/schemas/ObjectIdSchema"},"created_at":{"type":"string"},"img_urls":{"type":"array","items":{"type":"string"}},"likes":{"type":"integer","minimum":0},"message":{"type":"string"},"points":{"type":"integer","minimum":0},"tags":{"type":["array","null"],"items":{"$ref":"#/components/schemas/PostTag"}},"title":{"type":"string"},"user_id":{"type":"string"}}},"LikePost":{"type":"object","required":["post_id"],"properties":{"post_id":{"type":"string"}}},"LikeStatus":{"type":"string","description":"Generic like toggling function for any collection and object id\n\n","enum":["Like","Unlike"]},"ObjectIdSchema":{"type":"object","required":["$oid"],"properties":{"$oid":{"type":"string"}}},"PaginatedResponse_Comment":{"type":"object","required":["items"],"properties":{"items":{"type":"array","items":{"type":"object","required":["_id","user_id","post_id","message","created_at","likes","replies"],"properties":{"_id":{"$ref":"#/components/schemas/ObjectIdSchema"},"created_at":{"type":"string"},"likes":{"type":"integer","minimum":0},"message":{"type":"string"},"post_id":{"type":"string"},"replies":{"type":"array","items":{"$ref":"#/components/schemas/Reply"}},"user_id":{"type":"string"}}}}}},"PaginatedResponse_KamerlinkPost":{"type":"object","required":["items"],"properties":{"items":{"type":"array","items":{"type":"object","required":["_id","user_id","created_at","title","message","img_urls","likes","points"],"properties":{"_id":{"$ref":"#/components/schemas/ObjectIdSchema"},"created_at":{"type":"string"},"img_urls":{"type":"array","items":{"type":"string"}},"likes":{"type":"integer","minimum":0},"message":{"type":"string"},"points":{"type":"integer","minimum":0},"tags":{"type":["array","null"],"items":{"$ref":"#/components/schemas/PostTag"}},"title":{"type":"string"},"user_id":{"type":"string"}}}}}},"PaginatedResponse_PostTag":{"type":"object","required":["items"],"properties":{"items":{"type":"array","items":{"type":"object","required":["_id","tag_name","color","base_tag","uses"],"properties":{"_id":{"$ref":"#/components/schemas/ObjectIdSchema"},"base_tag":{"type":"boolean"},"color":{"type":"string"},"tag_name":{"type":"string"},"uses":{"type":"integer","minimum":0}}}}}},"PaginatedResponse_UserInfo":{"type":"object","required":["items"],"properties":{"items":{"type":"array","items":{"type":"object","required":["_id","nickname","role","is_validated","points"],"properties":{"_id":{"$ref":"#/components/schemas/ObjectIdSchema"},"is_validated":{"type":"boolean"},"nickname":{"type":"string"},"points":{"type":"integer","format":"int64"},"role":{"$ref":"#/components/schemas/Role"}}}}}},"PostDraft":{"type":"object","description":"\n\nPost request (creating a post)\n\n","required":["title","message","images"],"properties":{"images":{"type":"array","items":{"type":"array","items":{"type":"integer","format":"int32","minimum":0}},"contentMediaType":"application/octet-stream"},"message":{"type":"string"},"title":{"type":"string"}}},"PostResponse":{"type":"object","required":["post_id"],"properties":{"post_id":{"type":"string"}}},"PostTag":{"type":"object","required":["_id","tag_name","color","base_tag","uses"],"properties":{"_id":{"$ref":"#/components/schemas/ObjectIdSchema"},"base_tag":{"type":"boolean"},"color":{"type":"string"},"tag_name":{"type":"string"},"uses":{"type":"integer","minimum":0}}},"Reply":{"type":"object","required":["_id","user_id","message","created_at","likes"],"properties":{"_id":{"$ref":"#/components/schemas/ObjectIdSchema"},"created_at":{"type":"string"},"likes":{"type":"integer","minimum":0},"message":{"type":"string"},"user_id":{"type":"string"}}},"ReplyDraft":{"type":"object","required":["comment_id","message"],"properties":{"comment_id":{"type":"string"},"message":{"type":"string"}}},"ResponseGenericLike":{"type":"object","required":["status"],"properties":{"status":{"$ref":"#/components/schemas/LikeStatus"}}},"ResponseGivePoints":{"type":"object","properties":{"status":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/KamerlinkError"}]}}},"ResponseLikePost":{"type":"object","required":["status"],"properties":{"status":{"$ref":"#/components/schemas/LikeStatus"}}},"RetrieveBy":{"oneOf":[{"type":"null","description":"Retrieves items that have the same `_id` as the user\nwho is currently accessing the page.\n(Mostly used for the `User` type.)","default":null},{"type":"string","description":"Retrieves items with the given MongoDB ObjectId."},{"type":"null","description":"Retrieves all items, sorted by the number of likes (descending).","default":null},{"type":"null","description":"Retrieves all items, sorted by the number of points (descending).","default":null},{"type":"null","description":"Retrieves all items, sorted by creation date (`created_at` field, descending).","default":null},{"type":"null","description":"Retrieves the items that have uses set to the highest value. the values that have base_tag set to the highest value and then uses in descending order","default":null}],"description":"RetrieveBy can be either one of the fixed options (`\"_Self\"`, `\"MostLikes\"`, `\"MostPoints\"`, `\"MostRecent\"`) or any string (interpreted as an ID)."},"RetrievePaginated":{"type":"object","required":["type","page"],"properties":{"page":{"type":"integer","minimum":0},"type":{"$ref":"#/components/schemas/RetrieveBy"}}},"Role":{"type":"string","enum":["Student","Teacher","Admin"]},"UserInfo":{"type":"object","required":["_id","nickname","role","is_validated","points"],"properties":{"_id":{"$ref":"#/components/schemas/ObjectIdSchema"},"is_validated":{"type":"boolean"},"nickname":{"type":"string"},"points":{"type":"integer","format":"int64"},"role":{"$ref":"#/components/schemas/Role"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"security":[{"bearerAuth":[]}]}